+++ data.txt
@@ -1,1 +1,1 @@
-I put my keys on my bed
+I put my keys on the drawer
